Fix typos, bash formatting
There were some minor typos that I found.
Also in the web version of the documentation, any word that starts with the '$' symbol is highlighted as a variable. Many of the code examples used the '$' as the PS1 prompt, which caused the command to be highlighted as if it was a variable. Adding spaces after all of the PS1 prompts should fix this problem.