Stop timesync Tx stream when test (sim) finishes
Timesync is a very strange stream.  It is very much connected with
a latency stream, but should work the other way round.  TRex instance
that is a transmitting part (Tx) for the latency stream should be
a receiving part (Rx) for the timesync stream.  And vice-versa.
Tx for timesync is also called a master, while Rx is called a slave.

This patch does some refactoring of when and how the timesync Tx node
is initialized.  It has been moved to `TrexStatelessDpCore::add_stream`
where latency streams are created.  A timesync Tx node is created
quite similarly to a corresponding latency node.  This allows for it
to be similarly stopped and removed by `stop_traffic` method.

This patch is a follow up for what @hhaim mentioned
(codilime/trex-core#7) as a *restart issue*.