dpdk: add ability to specify driver devargs
DPDK allows you to specify runtime config options for a PCI device
by specifying devargs parameters in the white list. They can be used
to fine tune the driver perfromance or modify the behaviour without
any need to recompile a driver again. Provide the dpdk_devargs config
field to specify them in the TRex config file. This field is optional.

Signed-off-by: Alexander Kozyrev <akozyrev@mellanox.com>