ASTF: Use configured or discovered dest mac addresses
When the generator is used against a DUT that consists of a
router (gateway) and a device that performs Direct Server Return,
or any setup where traffic on a port is asymmetrical in terms of
layer2 counterpart, the source mac address of incoming packets
should not be used as destination for outgoing packets of a flow.
Instead the configured 'dest_mac' or discovered (ARP) mac address
of the configured 'default_gw' should be used.

By default, the original behavior of reflecting the MAC addresses
from the incoming packet is used. To enable the new behavior,
set the 'ip.dont_use_inbound_mac' to 1 in the server ASTFGlobalInfo
of the ASTFProfile.

Limitation: The MAC information is per port and hence this approach
will only work in case of a single gateway per port (but this is
not a new limitation).

Signed-off-by: David Verbeiren <david.verbeiren@tessares.net>