generate mlx5 driver config file as per environment
Hardcoded values in mlx_autoconf.h are not always correct
on different systems or with different OFED versions installed.
That may lead to compilation errors or wrong mlx5 configuration.

Generate this file according to the environment and get the
proper configuration for mlx5 DPDK driver. Take the script to
do this from DPDK and invoke it during ./b configure phase.

Signed-off-by: Alexander Kozyrev <akozyrev@nvidia.com>