Latency Measurement by HW timestamping with IEEE 1588 packets in Latency Stream
Using the 2-step mechanism of PTP protocol we send the accurate timestamp (t1)
of SYNC packet which is read from the NIC in a Follow Up packet rather than in
the same packet. Received timestamp (t2) is still taken on the first packet
(SYNC) itself. In this way with a combination of software(PTP protocol, DPDK
APIs) and HW support (leveraging IEEE 1588 support) the accuracy of latency
and jitter measurements can be improved.

Signed-off-by: Abhiram R N <arn@redhat.com>