fix 2 potential buffer overflows in DpdkTRexPortAttr
Both rte_eth_xstats_get and rte_eth_xstats_get_names do not include
xstats in their return values if the provided buffer is NULL. This
resulted in a case where if the number of xstats is greater than 0,
both DpdkTRexPortAttr::get_xstats_values and
DpdkTRexPortAttr::get_xstats_names would attempt to write past
the end of the vector provided to hold the return value. This was
observed to cause segfaults on both Intel and Broadcom NICs.

It may be possible to remove the do-while loop by passing in a valid
buffer in place of the NULL in both of these methods. However, that may
be less reliable since the do-while loop will continue attempting to get
the information until either it is successful or an error code is
returned.

Signed-off-by: Owen Hilyard <ohilyard@iol.unh.edu>