trex_astf_client: use proper to_json_str call
The profile object returned from ASTFProfile.load
can be of type ASTFProfileLight if the passed
"profile" is a path to a json profile file.

ASTFProfileLight has a different to_json_str
signature, and calling the method assuming
the returned profile object is of type
ASTFProfile will lead to failures in this
case.

Signed-off-by: bzalkilani <bzalkilani@gmail.com>